requirements-dev.txt
-r requirements.txt
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
black>=25.0,<26.0
ruff>=0.9,<1.0
django-debug-toolbar>=5.0,<6.0
factory-boy>=3.3,<4.0
